+++ internal/app.go
@@ -31,6 +31,7 @@ type Application struct {
 	natsConn      *nats.Conn
 	publisher     *natsclient.Publisher
 	subscriptions inboxapi.SubscriptionClient
+	settings      inboxapi.SettingsClient
 	feedRepo      *feed.Repo
 	feedService   *feed.Service
 	coreSDK       *coresdk.Client
@@ -104,9 +105,12 @@ func (a *Application) initDatabase() error {
 		conn = conn.Debug()
 	}
 
-	// TODO: Use real migrations intead of auto migrations from gorm
-	if err := conn.AutoMigrate(&feed.Item{}); err != nil {
-		return err
+	// TODO: Use real migrations instead of auto migrations from gorm
+	if err := conn.AutoMigrate(
+		&feed.Item{},
+		&feed.Settings{},
+	); err != nil {
+		return fmt.Errorf("automigrate: %w", err)
 	}
 
 	a.feedRepo = feed.NewRepo(conn)
@@ -143,6 +147,7 @@ func (a *Application) initInboxAPI() error {
 	}
 
 	a.subscriptions = inboxapi.NewSubscriptionClient(conn)
+	a.settings = inboxapi.NewSettingsClient(conn)
 
 	return nil
 }
@@ -154,7 +159,7 @@ func (a *Application) initCodeSDK() error {
 }
 
 func (a *Application) initServices() error {
-	a.feedService = feed.NewService(a.feedRepo, a.subscriptions, a.coreSDK)
+	a.feedService = feed.NewService(a.feedRepo, a.subscriptions, a.settings, a.coreSDK)
 
 	return nil
 }